verified domain connections refactored
