+++ api-js/src/verified-domains/objects/index.js
@@ -1,1 +1,2 @@
 export * from './verified-domain'
+export * from './verified-domain-connection'
